update user controller, service and repo

+++ service/user/user_service.go
@@ -2,12 +2,16 @@ package service
 
 import (
 	"context"
+	"errors"
 	"sync"
 
 	"github.com/Lukmanern/gost/domain/base"
 	"github.com/Lukmanern/gost/domain/entity"
 	"github.com/Lukmanern/gost/domain/model"
+	"github.com/Lukmanern/gost/internal/hash"
 	repository "github.com/Lukmanern/gost/repository/user"
+	"github.com/gofiber/fiber/v2"
+	"gorm.io/gorm"
 )
 
 type UserService interface {
@@ -39,16 +43,23 @@ func NewUserService() UserService {
 }
 
 func (service UserServiceImpl) Create(ctx context.Context, user model.UserCreate) (id int, err error) {
-	// Todo
-	// GetByEmail(ctx context.Context, email string) (user *model.UserResponse, err error)
-	// validasi dulu apakah email yang digunakan untuk regis sudah ada di database atau belum
+	userCheck, err := service.GetByEmail(ctx, user.Email)
+	if err == nil || userCheck != nil {
+		return 0, fiber.NewError(fiber.StatusBadRequest, "email has been used")
+	}
+
+	passwordHashed, hashErr := hash.Generate(user.Password)
+	if hashErr != nil {
+		return 0, errors.New("something failed while hashing data, please try again")
+	}
+
 	userEntity := entity.User{
 		Name:     user.Name,
 		Email:    user.Email,
-		Password: user.Password,
+		Password: passwordHashed,
 	}
-	userEntity.SetTimeAttributes()
-	// Todo : hash password
+	userEntity.SetTimes()
+
 	id, err = service.repository.Create(ctx, userEntity)
 	if err != nil {
 		return 0, err
@@ -59,11 +70,16 @@ func (service UserServiceImpl) Create(ctx context.Context, user model.UserCreate
 func (service UserServiceImpl) GetByID(ctx context.Context, id int) (user *model.UserResponse, err error) {
 	userEntity, err := service.repository.GetByID(ctx, id)
 	if err != nil {
+		if err == gorm.ErrRecordNotFound {
+			return nil, fiber.NewError(fiber.StatusNotFound, "data not found")
+		}
 		return nil, err
 	}
-	user.ID = userEntity.ID
-	user.Name = userEntity.Name
-	user.Email = userEntity.Email
+	user = &model.UserResponse{
+		ID:    userEntity.ID,
+		Name:  userEntity.Name,
+		Email: userEntity.Email,
+	}
 
 	return user, nil
 }
@@ -71,6 +87,9 @@ func (service UserServiceImpl) GetByID(ctx context.Context, id int) (user *model
 func (service UserServiceImpl) GetByEmail(ctx context.Context, email string) (user *model.UserResponse, err error) {
 	userEntity, err := service.repository.GetByEmail(ctx, email)
 	if err != nil {
+		if err == gorm.ErrRecordNotFound {
+			return nil, fiber.NewError(fiber.StatusNotFound, "data not found")
+		}
 		return nil, err
 	}
 	user = &model.UserResponse{
@@ -88,6 +107,7 @@ func (service UserServiceImpl) GetAll(ctx context.Context, filter base.RequestGe
 		return nil, 0, err
 	}
 
+	users = []model.UserResponse{}
 	for _, userEntity := range userEntities {
 		newUserResponse := model.UserResponse{
 			ID:    userEntity.ID,
@@ -102,10 +122,26 @@ func (service UserServiceImpl) GetAll(ctx context.Context, filter base.RequestGe
 }
 
 func (service UserServiceImpl) Update(ctx context.Context, user model.UserUpdate) (err error) {
+	isUserExist := func() bool {
+		getUser, getErr := service.GetByID(ctx, user.ID)
+		if getErr != nil {
+			return false
+		}
+		if getUser == nil {
+			return false
+		}
+
+		return true
+	}
+	if !isUserExist() {
+		return fiber.NewError(fiber.StatusNotFound, "data not found")
+	}
+
 	userEntity := entity.User{
 		ID:   user.ID,
 		Name: user.Name,
 	}
+	userEntity.SetUpdateTime()
 
 	err = service.repository.Update(ctx, userEntity)
 	if err != nil {
@@ -116,6 +152,21 @@ func (service UserServiceImpl) Update(ctx context.Context, user model.UserUpdate
 }
 
 func (service UserServiceImpl) Delete(ctx context.Context, id int) (err error) {
+	isUserExist := func() bool {
+		getUser, getErr := service.GetByID(ctx, id)
+		if getErr != nil {
+			return false
+		}
+		if getUser == nil {
+			return false
+		}
+
+		return true
+	}
+	if !isUserExist() {
+		return fiber.NewError(fiber.StatusNotFound, "data not found")
+	}
+
 	err = service.repository.Delete(ctx, id)
 	if err != nil {
 		return err
